Make wiz compile on 64-bit and 32-it architectures by using size_t instead uint/ulong for unsigned lengths in D.

+++ src/d/wiz/ast/number.d
@@ -6,9 +6,9 @@ import wiz.ast.lib;
 class Number : Expression
 {
     parse.Token _numberType;
-    ulong _value;
+    size_t _value;
 
-    this(parse.Token numberType, ulong value, compile.Location location)
+    this(parse.Token numberType, size_t value, compile.Location location)
     {
         super(location);
         _numberType = numberType;
